refactor(@angular-devkit/build-angular): cleanup stable architect API builder file names
