+++ packages/angular_devkit/build_angular/builders.json
@@ -2,50 +2,42 @@
   "$schema": "../architect/src/builders-schema.json",
   "builders": {
     "app-shell": {
-      "class": "./src/app-shell",
-      "implementation": "./src/app-shell/index2",
+      "implementation": "./src/app-shell",
       "schema": "./src/app-shell/schema.json",
       "description": "Build a server app and a browser app, then render the index.html and use it for the browser output."
     },
     "browser": {
-      "implementation": "./src/browser/index2",
-      "class": "./src/browser",
+      "implementation": "./src/browser",
       "schema": "./src/browser/schema.json",
       "description": "Build a browser app."
     },
     "dev-server": {
-      "class": "./src/dev-server",
-      "implementation": "./src/dev-server/index2",
+      "implementation": "./src/dev-server",
       "schema": "./src/dev-server/schema.json",
       "description": "Serve a browser app."
     },
     "extract-i18n": {
-      "class": "./src/extract-i18n",
-      "implementation": "./src/extract-i18n/index2",
+      "implementation": "./src/extract-i18n",
       "schema": "./src/extract-i18n/schema.json",
       "description": "Extract i18n strings from a browser app."
     },
     "karma": {
-      "class": "./src/karma",
-      "implementation": "./src/karma/index2",
+      "implementation": "./src/karma",
       "schema": "./src/karma/schema.json",
       "description": "Run Karma unit tests."
     },
     "protractor": {
-      "class": "./src/protractor",
-      "implementation": "./src/protractor/index2",
+      "implementation": "./src/protractor",
       "schema": "./src/protractor/schema.json",
       "description": "Run protractor over a dev server."
     },
     "tslint": {
-      "implementation": "./src/tslint/index2",
-      "class": "./src/tslint",
+      "implementation": "./src/tslint",
       "schema": "./src/tslint/schema.json",
       "description": "Run tslint over a TS project."
     },
     "server": {
-      "implementation": "./src/server/index2",
-      "class": "./src/server",
+      "implementation": "./src/server",
       "schema": "./src/server/schema.json",
       "description": "Build a server Angular application."
     }
