+++ packages/angular_devkit/build_angular/test/browser/allow-js_spec_large.ts
@@ -8,7 +8,7 @@
 import { Architect } from '@angular-devkit/architect/src/index2';
 import { join, normalize, relative, virtualFs } from '@angular-devkit/core';
 import { take, tap } from 'rxjs/operators';
-import { BrowserBuilderOutput } from '../../src/browser/index2';
+import { BrowserBuilderOutput } from '../../src/browser';
 import { createArchitect, host } from '../utils';
