+++ packages/angular_devkit/build_angular/test/browser/build-optimizer_spec_large.ts
@@ -7,7 +7,7 @@
  */
 import { Architect } from '@angular-devkit/architect/src/index2';
 import { join, normalize } from '@angular-devkit/core';
-import { BrowserBuilderOutput } from '../../src/browser/index2';
+import { BrowserBuilderOutput } from '../../src/browser';
 import { browserBuild, createArchitect, host } from '../utils';
