+++ packages/angular_devkit/build_angular/test/browser/index_spec_large.ts
@@ -8,7 +8,7 @@
 
 import { Architect } from '@angular-devkit/architect/src/index2';
 import { join, normalize, tags, virtualFs } from '@angular-devkit/core';
-import { BrowserBuilderOutput } from '../../src/browser/index2';
+import { BrowserBuilderOutput } from '../../src/browser';
 import { createArchitect, host } from '../utils';
