+++ packages/angular_devkit/build_angular/test/dev-server/public-host_spec_large.ts
@@ -7,7 +7,7 @@
  */
 import { Architect, BuilderRun } from '@angular-devkit/architect/src/index2';
 import fetch from 'node-fetch';  // tslint:disable-line:no-implicit-dependencies
-import { DevServerBuilderOutput } from '../../src/dev-server/index2';
+import { DevServerBuilderOutput } from '../../src/dev-server';
 import { createArchitect, host } from '../utils';
